Ejercicio Rectangulo-Perimetro {Error en la declaracion de tipo de dato :int->double}

+++ clases/src/com/krakedev/Rectangulo.java
@@ -9,8 +9,8 @@ public class Rectangulo {
 		return area;
 	}
 	
-	public int calcularPerimetro(){
-		int perimetro= (base*2)+(altura*2);
+	public double calcularPerimetro(){
+		double perimetro= (base*2)+(altura*2);
 		return perimetro;
 	}
 }
